- fixed bigDecimal for transaction - added tax consideration for stock sell

+++ src/main/java/io/stock/portfolio/backend/service/DividendService.java
@@ -115,7 +115,7 @@ public class DividendService {
                         .setDate(yahooSplit.getDate())
                         .setOperator(Operator.MULTIPLY)
                         //TODO will not work for reverse splits
-                        .setArgument(yahooSplit.getMultiplier().intValue())
+                        .setArgument(yahooSplit.getMultiplier())
         ).collect(Collectors.toSet());
 
         transactionRepository.saveAll(newSplitTransactions);
